requirements.txt
django
mysqlclient
pillow
django-ninja
django-cors-headers
requests
django-extensions

django-redis
django-celery-results
django-celery-beat


djangorestframework
djangorestframework-simplejwt
django-filter

celery==5.2.3
redis
gunicorn

django-ninja-jwt
python-jose[cryptography]

python-dotenv
black
pylint-django

google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0